Service skeletons for external data sources and config

+++ apps/backend/src/main/java/com/trendz/mvp/model/Trend.java
@@ -2,29 +2,15 @@ package com.trendz.mvp.model;
 
 public class Trend {
     private String name;
-    private double score;
+    private int score;
 
-    public Trend() {
-    }
-
-    public Trend(String name, double score) {
+    public Trend() {}
+    public Trend(String name, int score) {
         this.name = name;
         this.score = score;
     }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public double getScore() {
-        return score;
-    }
-
-    public void setScore(double score) {
-        this.score = score;
-    }
+    public String getName() { return name; }
+    public void setName(String name) { this.name = name; }
+    public int getScore() { return score; }
+    public void setScore(int score) { this.score = score; }
 }
